requirements-docker.txt
# optimized requirements for a container
aiosqlite==0.10.0
asyncpg==0.21.0
bcrypt==3.1.7
cffi==1.13.2
click==7.0
cryptography==2.8
databases==0.2.6
dnspython==1.16.0
email-validator==1.0.5
fastapi==0.62.0
h11==0.8.1
# httptools==0.0.13 # install via wheel to avoid gcc
idna==2.8
passlib==1.7.1
pycparser==2.19
pydantic==1.7.2
PyJWT==1.7.1
python-dotenv==0.15.0
python-multipart==0.0.5
six==1.15.0
SQLAlchemy==1.3.11
starlette==0.13.6
uvicorn==0.12.3
